Add .idea/vcs.xml and vercel.json files
Created new .idea/vcs.xml and vercel.json files. The .idea/vcs.xml is needed to correctly set up the Version Control System for the project in the IDE and the vercel.json file is required for properly routing API requests. These settings are required for the initial setup of the project.